Refactoring - added interface PhotoDao
