+++ src/main/java/com/photostudio/ServiceLocator.java
@@ -1,7 +1,8 @@
 package com.photostudio;
 
 
-import com.photostudio.dao.LocalDiskPhotoDao;
+import com.photostudio.dao.PhotoDao;
+import com.photostudio.dao.file.LocalDiskPhotoDao;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.photostudio.dao.OrderDao;
 import com.photostudio.dao.UserDao;
@@ -40,8 +41,8 @@ public class ServiceLocator {
         OrderDao orderDao = new JdbcOrderDao(dataSource);
         register(OrderDao.class, orderDao);
 
-        LocalDiskPhotoDao photoDiskDao = new LocalDiskPhotoDao(properties.getProperty("dir.photo"));
-        register(LocalDiskPhotoDao.class, photoDiskDao);
+        PhotoDao photoDiskDao = new LocalDiskPhotoDao(properties.getProperty("dir.photo"));
+        register(PhotoDao.class, photoDiskDao);
 
         OrderService orderService = new DefaultOrderService();
         register(OrderService.class, orderService);
