+++ src/main/java/com/photostudio/dao/PhotoDao.java
@@ -1,0 +1,6 @@
+package com.photostudio.dao;
+
+
+public interface PhotoDao {
+    void deleteByOrder(long orderId);
+}
