+++ src/main/java/com/photostudio/service/impl/DefaultOrderService.java
@@ -1,7 +1,8 @@
 package com.photostudio.service.impl;
 
 import com.photostudio.ServiceLocator;
-import com.photostudio.dao.LocalDiskPhotoDao;
+import com.photostudio.dao.PhotoDao;
+import com.photostudio.dao.file.LocalDiskPhotoDao;
 import com.photostudio.dao.OrderDao;
 import com.photostudio.entity.order.FilterParameters;
 import com.photostudio.entity.order.Order;
@@ -14,7 +15,7 @@ import java.util.List;
 public class DefaultOrderService implements OrderService {
     private final Logger LOG = LoggerFactory.getLogger(getClass());
     private OrderDao orderDao = ServiceLocator.getService(OrderDao.class);
-    private LocalDiskPhotoDao photoDao = ServiceLocator.getService(LocalDiskPhotoDao.class);
+    private PhotoDao photoDao = ServiceLocator.getService(PhotoDao.class);
 
     @Override
     public List<Order> getAll() {
